refactor: implement absolute import for components and rename files
