+++ src/components/TruckForm.tsx
@@ -3,9 +3,9 @@ import { useEffect, useState } from "react";
 import api from "../api/axios";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { truckSchema, TruckData } from "../schemas/truckSchema";
-import Input from "./input";
-import { Button } from "./button";
-import Loader from "./loader";
+import Input from "./Input";
+import { Button } from "./Button";
+import Loader from "./Loader";
 
 type Props = {
   onSubmit: (data: TruckData) => void;
