+++ src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client'
 import { QueryClientProvider, QueryClient } from '@tanstack/react-query'
 import { RouterProvider } from 'react-router';
 import { router } from './routes/routes';
-import ToastController from './components/toast';
+import ToastController from './components/ToastController';
 
 const queryClient = new QueryClient();
