+++ src/pages/FirstAcessChangeForm.tsx
@@ -1,54 +1,53 @@
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import api from "../api/axios";
-import Input from "../components/input";
-import { Button } from "../components/button";
-import Loader from "../components/loader";
+import Input from "@components/Input";
+import Loader from "@components/Loader";
+import { Button } from "@components/Button";
 import { toast } from "react-toastify";
+import { zodResolver } from "@hookform/resolvers/zod";
 import { Navigate, useNavigate } from "react-router-dom";
-import { z } from "zod";
-import { ResetConfirmData as FirstAccessData, resetConfirmSchema as firstAccessSchema } from "../schemas/loginSchema";
-import { useEffect, useState } from "react";
+import {
+  ResetConfirmData as FirstAccessData,
+  resetConfirmSchema as firstAccessSchema,
+} from "@schemas/loginSchema";
+import { useForm } from "react-hook-form";
+import { useState } from "react";
+import api from "src/api/axios";
 
 export default function FirstAccessPasswordChange() {
   const {
     register,
     handleSubmit,
     formState: { errors, isSubmitting },
-    setValue
   } = useForm<FirstAccessData>({
     resolver: zodResolver(firstAccessSchema),
   });
 
   const navigate = useNavigate();
 
-  const [confirmPassword, setConfirmPassword]= useState("");
-  const [confirmPasswordError, setConfirmPasswordError] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
+  const [confirmPasswordError] = useState("");
 
   async function handleFirstAccessChange(data: FirstAccessData) {
     if (confirmPassword !== data.newPassword) {
       toast.error("As senhas não coincidem!");
       return;
-    } 
+    }
     try {
-
       const userData = await api.get("/users/me");
 
-      if(!userData.data.passwordChangeRequired){
-          return <Navigate to="/main" replace />;
+      if (!userData.data.passwordChangeRequired) {
+        return <Navigate to="/main" replace />;
       }
 
-      const token = localStorage.getItem("accessToken");
-
+      //const token = localStorage.getItem("accessToken");
       const res = await api.post("/auth/password/first-access-change", data);
- 
+
       if (res.status === 200) {
         toast.success("Senha alterada com sucesso!");
         navigate("/login");
       } else {
         toast.error("Erro ao alterar a senha.");
       }
-    } catch (err) {
+    } catch {
       toast.error("Erro inesperado ao trocar a senha.");
     }
   }
@@ -66,25 +65,25 @@ export default function FirstAccessPasswordChange() {
         </h1>
 
         <Input
-        text="Senha"
-        type="password"
-        id="newPassword"
-        placeholder="Insira a senha do usuário..."
-        register={register}
-        error={errors.newPassword?.message}
-      />
+          text="Senha"
+          type="password"
+          id="newPassword"
+          placeholder="Insira a senha do usuário..."
+          register={register}
+          error={errors.newPassword?.message}
+        />
 
-      <Input
-        text="Confirmar Senha"
-        type="password"
-        id="passwordConfirm"
-        placeholder="Confirme a senha do usuário  ..."
-        value = {confirmPassword}
-        onChange={(e) => {
-          setConfirmPassword(e.target.value)
-        }}
-        error={confirmPasswordError}
-      />
+        <Input
+          text="Confirmar Senha"
+          type="password"
+          id="passwordConfirm"
+          placeholder="Confirme a senha do usuário  ..."
+          value={confirmPassword}
+          onChange={(e) => {
+            setConfirmPassword(e.target.value);
+          }}
+          error={confirmPasswordError}
+        />
 
         <Button
           type="submit"
