+++ src/pages/Home.tsx
@@ -1,6 +1,4 @@
-import { Link } from "react-router-dom";
-import { Button } from "../components/button";
-import Header from "../components/header";
+import Header from "@components/Header";
 
 export default function Home() {
   return (
