+++ src/pages/Login.tsx
@@ -1,4 +1,5 @@
-import Form from "../components/form/form"
+import Form from "@components/LoginForm"
+
 export default function Login() {
   return (
     <div className='flex justify-center items-center h-screen bg-[#EFEAE6]'>
