--- tsconfig.app.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/api/axios.ts","./src/components/TruckForm.tsx","./src/components/button.tsx","./src/components/header.tsx","./src/components/input.tsx","./src/components/link.tsx","./src/components/loader.tsx","./src/components/select.tsx","./src/components/toast.tsx","./src/components/form/form.tsx","./src/components/form/input.tsx","./src/lib/utils.ts","./src/pages/FirstAcessChangeForm.tsx","./src/pages/Home.tsx","./src/pages/Login.tsx","./src/pages/Main.tsx","./src/pages/PasswordResetConfirmForm.tsx","./src/pages/PasswordResetRequestForm.tsx","./src/pages/RegisterForm.tsx","./src/pages/TruckList.tsx","./src/routes/adminRoutes.tsx","./src/routes/managerRoutes.tsx","./src/routes/protectedRoutes.tsx","./src/routes/routes.tsx","./src/schemas/loginSchema.ts","./src/schemas/truckSchema.ts"],"errors":true,"version":"5.8.3"}
